refactor: use project metadata submit hook
